Created high level block diagram of server hardware/software
I think this summarises what we discussed in our first meeting.
I may have gotten carried away.

The other approach suggested was to use some sort of custom hardware,
with a wireless link to a hand held device.

Making this diagram reinforces my belief that we shouldn't do that.
We'd have to implement the same blue stuff as well as all of the green stuff.
Also there would be more red stuff.

That's my technical explanation :P

Sam Moore