More work on webserver test
It can send a text file straight to a client.

Something wierd happens.

If you use chrome to look at a html file, it displays fine.
If you use iceweasel (ie: firefox) to look at the same file, it just shows as plain text.

Using netcat to test, a real webserver just dumps html files the same way.
So what is (not) happening to make iceweasel fail to recognise it as html?