Test webserver with minimalist JavaScript
I'm not sure if we want to use jQuery, but it works for testing.

I think this webserver is a good proof of concept.
I will get it to set or read a GPIO pin on the raspberry pi.

Then I can try the same thing with a CGI script under apache2.