Fixed binary file access
Really dumb mistakes.

If we can open the file for reading and writing, and remove the need to open/close
in both threads each time it is accessed, the efficiency might be better.

We'd still have to close the file to save changes to disk though,
might be a problem if the program crashes.