Added restart functionality to run.sh
Essentially if there is a non zero (error) exit, it will restart the program.
If it gets a massive number of fatal errors it will eventually give up.

We should do something similar to this.
Perhaps receive a certain number of fatal errors in a time interval to give up?

Perhaps make the software not have Fatal errors :P

Making sure cleanup code is called might be important.

Also we really need the errors to show up in the GUI.

Anyway, this is useful for now because it lets me do things like this:
http://192.168.1.10/api/pin?type=gpo&num=13

To find the non working pins (eg: GPIO13) without restarting the server a billion times.