MAJOR refactoring of Sensors code
Will probably get in trouble for this...

Abstracted out the threaded sensor control logic from the actual initialisation/reading/cleanup of sensors.
Adding a new sensor to the program still isn't just a one line process, but it's a lot more general than before.

Added test sensors for the strain gauges, total CPU usage (because that seemed interesting)

Went a bit insane with power and added a "Piped" sensor type which lets people run external programs for the sensors :S
(As long as it prints doubles to stdout and is unbuffered it will work).