Add UserCake
We might be able to build on this to make a user management system.

UserCake Version: 2.0.2
http://usercake.com

This commit includes the original UserCake source.
Many files will need to be removed and/or rewritten.

UserCake uses MySQL for storing user information.
There is a C API for interfacing with MySQL. Hopefully that will work.