Begin modifying UserCake
Create admin upload page for uploading / purging users. Incomplete.
Note that this page needs to be secured somehow.

Realised that we probably need to add the UserCake install scripts as well, so we can
customise the database and other things when it is installed.

Modified something in the server too apparently.
I love it when I forget what I actually did...